Merge 165615 - Flash Mac: Fix escape to exit fullscreen
On tmz.com, pressing escape to exit fullscreen Flash would cause the currently playing movie to close.

The problem was that the fullscreen window was closing on a key down event and the key up event was going to the parent site.

My fix is to suppress the key up event on the parent view.

BUG=155492


Review URL: https://chromiumcodereview.appspot.com/11188027

TBR=sail@chromium.org
Review URL: https://codereview.chromium.org/11275152

git-svn-id: https://src.chromium.org/chrome/branches/1312/src@165946 4ff67af0-8c30-449e-8e8b-ad334ec8d88c