Merge 165424 - Free the dibsection (HBITMAP) in PlatformBitmap's destructor, so we don't leak!
This adds an extra slot to PlatformBitmap, which is implemented for all platforms. To
do this simply, an intptr_t field is added (platform_extra_) which is initialized to 0.
At the moment, only the windows port uses this, but there is no cost to the other ports.

BUG=158931
Review URL: https://codereview.chromium.org/11293036

TBR=reed@google.com
Review URL: https://codereview.chromium.org/11365092

git-svn-id: https://src.chromium.org/chrome/branches/1312/src@165954 4ff67af0-8c30-449e-8e8b-ad334ec8d88c