Merge 165526 - Fix version discovery for Interfaces, Structs
Now that we are looking at including structures and interfaces pointers in the
version graph, we need to fix remove some simplifying  assumptions, and correct
the versioning calculation.

Includes minor cleanup to how we call 'Main' to make it more compliant as
well as catching exceptions so that we show the commandline options that caused
that exception.

BUG=154127,158980
TBR=sehr@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11263062

TBR=noelallen@chromium.org
Review URL: https://codereview.chromium.org/11359216

git-svn-id: https://src.chromium.org/chrome/branches/1312/src@167729 4ff67af0-8c30-449e-8e8b-ad334ec8d88c