Merge 168093 - Don't wait endlessly for flushing the batched queries sent to the GPU driver during H/W decode
We currently have a loop in the DXVA decoder on Windows 7 and above where we flush
the queued D3D command bufers and wait for the operation to complete. This causes an 
infinite loop on certain multicore machines due to a bug in the IDirect3DQuery9::GetData
call not returning the correct result. This seems to be a very old bug dating back to April 2008
http://us.generation-nt.com/idirect3dquery9-getdata-fails-return-ok-multi-core-cpus-help-26986872.html

Fix for now is to have an upper limit of 10 iterations while we wait for the flush to complete. ld

BUG=149139
R=apatrick
Review URL: https://chromiumcodereview.appspot.com/11369229

TBR=ananta@chromium.org
Review URL: https://codereview.chromium.org/11414042

git-svn-id: https://src.chromium.org/chrome/branches/1312/src@168335 4ff67af0-8c30-449e-8e8b-ad334ec8d88c